desde el fork, commit: a00ff9292a6c9932ea7818864178d8e6fe767762

+++ src/main/java/ucu/ast/Stmt.java
@@ -42,7 +42,7 @@ public abstract class Stmt extends AstNode {
 		if (obj instanceof Integer){
 			return ((Integer) obj).doubleValue();
 		}else if (obj instanceof Double){
-			return ((Double) obj).doubleValue();
+			return (Double) obj;
 		}else{
 			throw new IllegalStateException(unparse);
 		}
